Added discord and slack integration

+++ php/functions/account.php
@@ -27,8 +27,7 @@ class Account
 			return array('success'=>false, 'output'=>"This address is already in use.");
 		}
 
-		$usercheck = db_query("SELECT address FROM verification WHERE address = ?", array($address));
-
+		$usercheck = db_query("SELECT address FROM verification WHERE address = ? AND user_id = ?", array($address, $_SESSION['id']));
 		if (count($usercheck))
 		{
 			return array('success'=>false, 'output'=>"Verification code has already sent.");
@@ -61,7 +60,7 @@ class Account
 			error_log("Failed to send verification code to $address. " . $mail->ErrorInfo);
 			return array("success"=>false, "output"=>"Failed to send message.");
 		} else {
-			db_query("INSERT INTO verification(code, address) VALUES (?, ?)", array($code, $address));
+			db_query("INSERT INTO verification(code, address, user_id) VALUES (?, ?, ?)", array($code, $address, $_SESSION['id']));
 			return array("success"=>true);
 		}
 	}
@@ -75,6 +74,12 @@ class Account
 			return array('success'=>false, 'output'=>'Invalid name.');
 		}
 
+		$namecheck = db_query("SELECT name FROM account WHERE name = ? AND user_id = ?", array($name, $_SESSION['id']));
+		if (count($namecheck))
+		{
+			return array('success'=>false, 'output'=>"This name is already in use.");
+		}
+
 		if ($form['type'] === 'Phone' || $form['type'] === 'Email')
 		{
 			if ($form['type'] === 'Phone')
@@ -85,44 +90,59 @@ class Account
 				$address = trim($form['address']);
 			}
 
-			$usercheck = db_query("SELECT address FROM verification WHERE address = ? AND code = ?", array($address, $form['verification']));
-			if (!count($usercheck))
+			$codecheck = db_query("SELECT address FROM verification WHERE address = ? AND code = ?", array($address, $form['verification']));
+			if (!count($codecheck))
 			{
 				return array('success'=>false, 'output'=>"Invalid verification code.");
 			}
 
-			$namecheck = db_query("SELECT name FROM account WHERE name = ? AND user_id = ?", array($name, $_SESSION['id']));
-			if (count($namecheck))
-			{
-				return array('success'=>false, 'output'=>"This name is already in use.");
-			}
-
 			db_query("INSERT INTO account(user_id, name, address, type) VALUES (?, ?, ?, ?)", array($_SESSION['id'], $name, $address, strtolower($form['type'])));
-			db_query("DELETE FROM verification WHERE address = ?", array($address));
+			db_query("DELETE FROM verification WHERE address = ? AND user_id = ?", array($address, $_SESSION['id']));
 
-			$icon = "fa-question-circle-o";
-			switch (strtolower($form['type']))
-			{
-				case 'phone':
-					$icon = 'fa-mobile';
-					break;
-				case 'email':
-					$icon = 'fa-envelope-o';
-					break;
-				case 'slack':
-					$icon = 'fa-slack';
-					break;
-				case 'discord':
-					$icon = 'fa-user-circle';
-					break;
-			}
-
-			return array("success"=>true, "output"=>array("name"=>$name, "address"=>$address, "icon"=>"$icon"));
 		}
 		elseif ($form['type'] === "Slack" || $form['type'] === "Discord")
 		{
-			error_log('slack/discord');
+			$hook_format = array(
+				"Slack"=>"https://hooks.slack.com/services",
+				"Discord"=>"https://discordapp.com/api/webhooks"
+			);
+
+			// check if valid webhook format
+			if (substr($form['webhook'], 0, strlen($hook_format[$form['type']])) !== $hook_format[$form['type']])
+			{
+				return array('success'=>false, 'output'=>"Invalid webhook format.");
+			}
+
+			if (strlen($form['channel']) < 8)
+			{
+				return array('success'=>false, 'output'=>"Invalid channel ID.");
+			}
+
+			$extra = json_encode(array('channel'=>$form['channel']));
+
+			db_query("INSERT INTO account(user_id, name, address, type, extra) VALUES (?, ?, ?, ?, ?)", array($_SESSION['id'], $name, $form['webhook'], strtolower($form['type']), $extra));
 		}
+
+		$icon = "fa-question-circle-o";
+		switch (strtolower($form['type']))
+		{
+			case 'phone':
+				$icon = 'fa-mobile';
+				break;
+			case 'email':
+				$icon = 'fa-envelope-o';
+				break;
+			case 'slack':
+				$icon = 'fa-slack';
+				break;
+			case 'discord':
+				$icon = 'fa-user-circle';
+				break;
+		}
+
+		$account = db_query("SELECT * FROM account WHERE name = ? AND user_id = ?", array($name, $_SESSION['id']));
+
+		return array("success"=>true, "output"=>array("id"=>$account['id'], "name"=>$name, "icon"=>"$icon"));
 	}
 }
